HBX-1670: Move class 'org.hibernate.tool.hbmlint.detector.EntityModelDetector' to 'org.hibernate.tool.internal.export.lint.EntityModelDetector'
