+++ orm/src/main/java/org/hibernate/tool/internal/export/lint/BadCachingDetector.java
@@ -5,7 +5,6 @@ import org.hibernate.mapping.PersistentClass;
 import org.hibernate.mapping.Property;
 import org.hibernate.mapping.Value;
 import org.hibernate.tool.hbm2x.visitor.EntityNameFromValueVisitor;
-import org.hibernate.tool.hbmlint.detector.EntityModelDetector;
 
 public class BadCachingDetector extends EntityModelDetector {
